First versions
The Ticketing System is a Java program that provides a simple user interface for purchasing airplane tickets. It allows users to enter passenger details such as name, flight information, departure and arrival airports, departure date, and seat number. The program utilizes object-oriented programming concepts and provides classes such as AirplaneTicket, Person, and LinkedListNode for modeling the ticketing system. The user interface offers menu options to purchase tickets and gracefully exit the program. This project serves as a starting point for building more complex ticketing systems or can be customized to fit specific requirements.